pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "readme-doc-healer"
version = "0.1.0"
description = "MCP server that diagnoses legacy API documentation gaps, assembles context for healing, and audits live ReadMe projects."
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "pyyaml>=6.0",
    "jsonref>=1.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "jsonschema>=4.20",
    "openapi-spec-validator>=0.7",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "rich>=13.0",
]

[project.scripts]
readme-doc-healer = "readme_doc_healer.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/readme_doc_healer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
